requirements.txt
telethon
python-telegram-bot
cryptography
python-dotenv